requirements.txt
tkinter
pandas
openpyxl
python-docx
dash
dash-cytoscape